Add functionality to add contacts and display them in the contact list
This commit adds the ability to add contacts to a user's contact list and display them in the contact list. The `addContacts` function in `FireDatabase` is used to add a contact to the current user's `my_users` array in Firestore. The contact list is then updated to display the newly added contact.